Move audio player to common/
Start putting common functionality in one place
