+++ src/common/audio.js
