--- src/survey/components/audio.css
@@ -1,10 +1,0 @@
-.audio {
-  width: 120px;
-  transition: all 3s;
-}
-
-.audio__button {
-  display: block;
-  border: none;
-  background-color: transparent;
-}
